Fix int32/int64 indexing in CUDA kernels for large tensors
Replaced int-based indexing with int64_t casts in CUDA kernels across moe_cuda_ops.cu, transformer_ops.cu, fused_grad_clip.cu, and fused_loss.cu to ensure correct memory access for large tensors. Also added __CUDA_ARCH__ checks for bf16 kernels in transformer_ops.cu. These changes improve compatibility and correctness when handling large models or batch sizes.

+++ Src/Main_Scripts/training/fused_loss.cu
@@ -65,7 +65,7 @@ __global__ void __launch_bounds__(256, 4) fused_cross_entropy_single_pass(
 
   const int64_t label = labels[token_idx];
 
-  const float *logit_row = logits + (size_t)token_idx * vocab_size;
+  const float *logit_row = logits + (int64_t)token_idx * vocab_size;
   const float weight =
       (loss_weights != nullptr) ? loss_weights[token_idx] : 1.0f;
